Fix edge cases in selective execution (#4140)
* Fixes handling of private tasks and added tests for that and for
renaming tasks. As private tasks and overridden tasks are assigned a
`.super.xyz` suffix during planning, we need to ensure we use the
`Terminal#render` label which includes this suffix rather than the
`.ctx.segments.render` label which does not

* Ensure we preserve all watches during `--watch`, even if some portion
of the build graph was skipped due to selective execution. This ensures
that inputs that were skipped can continue to be watched and still
trigger downstream changes later

* We should make sure in `selective.resolve` that we do not call
`resolve` on the outcome of `diffMetadata`, as that may include private
tasks that cannot be resolved externally. Instead we should only call
`resolve` when there is no diff to be had, in which case we need to take
the input selector and resolve it to concrete task names to display