Universal availability check for an interactive console with JDK >= 22 support (#4178)
JDK 22 introduced some changes to the method `System.console()`:

> In JDK 22, System.console() has been changed to return a Console with
enhanced editing features that improve the experience of programs that
use the Console API. In addition, System.console() now returns a Console
object when the standard streams are redirected or connected to a
virtual terminal. Prior to JDK 22, System.console() instead returned
null for these cases. This change may impact code that checks the return
from System.console() to test if the JVM is connected to a terminal. If
required, the -Djdk.console=java.base flag will restore the old behavior
where the console is only returned when it is connected to a terminal.
Starting JDK 22, one could also use the new Console.isTerminal() method
to test if the console is connected to a terminal.

Mill assumes the old behavior of `System.console` and use
`System.console() != null` to check if we have interactive console,
causing issue like https://github.com/com-lihaoyi/mill/issues/4161.

This PR added universal availability check for an interactive console,
compatible with JDK < 22 and later.

SBT made similar changes in https://github.com/sbt/sbt/pull/7843.

Closes #4161

---------

Co-authored-by: Nikolay Artamonov <n.artamonov@sigma-it.ru>