Working towards android support (#4169)
## Part of Android support

### Relevant issues: 

- https://github.com/com-lihaoyi/mill/issues/3868
- https://github.com/com-lihaoyi/mill/issues/3550

### Outline

- Follow android studio + gradle standard for directory structure
- Unit test support (local runs - no android device needed)
- Integration test support (android tests, needs an android
device/emulator)

### New features

- Ability to install an apk in the default available android device
(tested against AVD)
- Ability to run all integration tests with the install command (android
manifest with instrumentation is needed, this needs to be a
functionality of integration tests support in the future instead)
- Classpath resolution (runtimeDependencies) for packaging with
dependencies and user's source code (tested against integration test and
running the android app itself via the emulator)
- Special classpath resolution and packaging for android tests
(integration tests)
- Mill integrates with avdmanager, adb and emulator to automate the run
of integration tests by creating virtual devices and controlling the
emulator (start, stop). Examples are added in documentation of both
kotlin and java app. This includes the ability to use different
emulators for different apps, by specifying identifiers and ports,
though this use case should be rare in app development lifecycles.
- Github actions & mill support to run integration tests in (headless)
emulator & differentiate from local mode (windowed emulator + default
gpu settings)

### Github actions

A sample repo to demonstrate what's needed to run the emulator
https://github.com/vaslabs-ltd/test-android-emulator

In the end what slowed me down was that for github actions, the AVD home
env var was needed, which wasn't obvious until I've tested it in
isolation . All the flags and kvm configs are also needed and it's what
https://github.com/ReactiveCircus/android-emulator-runner does

### Stubs and missing features

An effort was made to standardise the structure faster, so follow up PRs
will not use a non-standard structure. Things that are working but not
"production ready" are:

1. Integration tests debug apk: This inherits the current apk
construction, debug functionality needs to be added and separated from
the release androidApk task
2. No testOnly for integration tests: this is easy to do, but was not
added in the interest of not increasing the size of the PR and the
testing surface
3. Classpath resolution: For kotlin, jvm specific dependencies are
removed in a hardcoded way.
4. Desugaring: This shows as a warning in the android tests apk
creation, while the flow works, the d8 will need to be configured to
desugar annotations properly.
5. Emulator CI mode only recognises github actions

### Demo

#### Running android tests


https://github.com/user-attachments/assets/0cb47ac7-32ff-4d5c-a678-11cd18807b10

---------

Co-authored-by: irodotos7 <irodotos@vaslabs.io>