Check if rt.jar is still necessary (#4206)
Fixes https://github.com/com-lihaoyi/mill/issues/4164

It's been around since we used to use Ammonite to compile scripts, and
IIRC was necessary to make the in-memory Scala compiler work. We no
longer go through that code path, so maybe we can get rid of it? AFAIK
we now only use the Scala compiler through Zinc which manages this stuff
itself