Android: add aar dependencies support, create proper build pipeline, add Jetpack Compose example (#4188)
This PR changes a lot of things in the Android build support (mainly
focusing on creating proper build pipeline). In details:

* `aar` dependencies processing
* Manifest merging
* Setting min/target/compile SDK properties in the build script
properties instead of manifest
* Proper work with `aapt2` at the `compile`/`link`stages
* R classes generation for the libraries (hacky though)
* Proguard definitions collection. NB: there is no support of
obfuscation in this PR, Proguard here is supported only for the main DEX
packing.
* Clear separation between debug and release signing configs (release
key shouldn't be a part of the required setup; user should be able to
bootstrap project quickly and create release key only when it is really
needed).
* Changes to the Android build script APIs
* Introduction of the debug vs release flags for `aapt2` and dexing
* Plus it adds Jetpack Compose sample

This PR **doesn't add yet** support of the transitive Android modules,
this is a work to be done in the future.

The code quality is maybe meh, but I expect many pieces here to be
changed in the future to be better and have wider support of the
different build options/flags.

The most important part this PR does is mentioned above: it brings a
proper pipeline. I added many flags/comments for the future work,
allowing contributors to focus on the specific parts.

One thing to mention as well is that in the current state the
incremental builds will have terrible performance. This is the area
which requires a massive time investment to be on-par with AGP to
support only processing of the files in the pipeline which really
changed. Because of this the output types of certain tasks will probably
change in the future or they will be split into more granular tasks.

I would like also to mention other things which should be improved:

* Things like `def bundleToolVersion = "1.17.2"` should be pre-defined
in the Android SDK module and they shouldn't be required for writing
build script, because:
* in this particular example bundle is not needed for APK modules or lib
modules;
* these things are really technical and it is not easy to get versions
list without knowing artifact coordinates
* ideally Mill should be able to fetch latest versions of such tooling
by itself
* Project structure in the Android examples created so far is very
different from project structure created by Android Studio, this may
bring additional complexity migrating existing projects. Maybe it is
worth to create something similar to `MavenModule`?
* Successful APK/bundle compilation doesn't guarantee that everything is
alright, app may crash at runtime because something wasn't packed
correctly. It is needed to add integration tests support on the Android
emulator as a part of CI pipeline
* **IMPORTANT**: Release key tasks for getting key alias/pass/store pass
are leaking sensitive credentials, because outputs will be serialized to
JSON. Sensitive data should never be exposed (let it be disk or console
output).
* Maybe it is worth to create `mill.androidlib`?
* I think it is not worth to focus on Java samples right now, given that
Kotlin is the preferred language for the Android development.
* I don't thing creating many traits like `AndroidAppBundle` is a good
idea, because it is just a way of packaging. AGP has only 2 main
entrypoints: Application and Library, the rest is their configuration
options. Bazel also has application and library as main entrypoints for
rules (see https://github.com/bazelbuild/rules_android/tree/main/rules).
* Default build tools version should be inferred from `compileSdk`
version (e.g `compileSdk` is 35, then `buildToolsVersion` should be
`35.0.0` by default), but in the current layout it is not possible to do
implicitly, so both versions should be specified.

Co-authored-by: 0xnm <0xnm@users.noreply.github.com>