Part of Android Studio Integration: Fix apk installation failure (#4313)
### Motivation 
The following small PR is part of integrating android studio with mill,
so it is focused on developer experience. This was already mentioned as
a TODO in the androidDebugKeystore method. The screenshot below is with
an attempt to hook together android studio and mill and deleting the out
directory, then trying to run the bsp test (work pending)


![image](https://github.com/user-attachments/assets/90688bcc-b919-41ee-bd43-6a65436d7485)

So far, the debug keystore, is stored in the task destination directory,
which makes it ephemeral, even for a local development device, as out
directory can be wiped frequently.

### Advice needed

We followed a similar approach that gradle and android studio use,
excluding the lock file. So I'm marking this PR as draft and asking for
opinions on:
1. should we handle multi-process access to the debug keystore (i.e. use
a file lock)?
2. is it a problem that we break sandbox (we already do with android
emulator) and if it is, should we maintain the current functionality for
mill sandbox?


### Context

As part of Android studio integration, we are trying to make the bsp
integration work for Mill Android module. With some crude adjustments,
some functionality starts working (e.g. change bspCompileClasspath to be
in sync with compile classpath) and then configure the testTask to work
properly with test results etc (work pending).

The aim is to find small things that are not working and send small PRs
instead of an uber PR in the end.


This is co-authored with @irodotos7

### Update

For now, I've kept the keystore in debug destination, but its original
location is the global one (so it's copied over), so the persistent is
still working, but now it is also idempotent (assuming the original file
in global position is not changed) . Not sure if this is actually
needed, I think just having the global dest is much simpler but let's
see how it goes and how this evolves