Part of Android Support: screenshot testing (#4411)
# Background

Screenshot testing is a way to test UIs and in Android's case, typically
without the need of running an emulator. This is achieved in recent
developments with the usage of Compose and Previews (the same
functionality that generates previews in android studio).

The typical flow is:

1. A base line is generated, with screenshots being committed into the
application's repo
2. Regression tests subsequently, compare the screenshots
3. For new UIs that devs decide to add coverage or agreed changes,
baseline is updated

Further Reading:
https://developer.android.com/studio/preview/compose-screenshot-testing

# Provided in this PR

This PR, while it has in mind the full flow support, only provides the
means to generate screenshots and run diffs with the same tools that AGP
is using. At the same time, we keep the test task compatibility with
Mill so we don't need to implement any custom runner tasks.

## Generating screenshots

Using the same tool as AGP, the
[compose-preview-renderer](https://android.googlesource.com/platform/tools/base/+/61923408e5f7dc20f0840844597f9dde17453a0f/preview/screenshot/screenshot-test-gradle-plugin/src/main/java/com/android/compose/screenshot/tasks/PreviewRenderWorkAction.kt),
which has undocumented cli arguments. However, these can be found by
running the respective gradle task in a sample Android studio project

```bash
./gradlew  updateDebugScreenshotTest
```

Then, given the generated data, the PR worked towards generating the
file computing classpaths and other needed files similarly

![CLI arguments generated by Gradle for the preview
tool](https://github.com/user-attachments/assets/f358a93c-5f09-4000-b8a3-c814bb2de4e7)

Currently these screenshots are ephemeral, and this PR is not concerned
by adding support for version control on the screenshot base. The plan
is for a future PR to make it easier for the users to customise the
behaviour (e.g. choose the path)

## Testing screenshots

AGP uses the
[screenshot-validation-junit-engine](https://android.googlesource.com/platform/tools/base/+/61923408e5f7dc20f0840844597f9dde17453a0f/preview/screenshot/screenshot-validation-junit-engine).
This PR provides similar functionality but hooking into Mill's testing
infrastructure instead of Gradle.

## Example App

The example App was set up initially with AGP and then a mill version
was constructed with 2 differences:

1. Only 1 screen configuration is provided for the simplicity of this
first attempt
2. Although compose preview is technically working, it is only kept in
the screenTest module, again for the simplicity of this PR and not doing
any changes to the compose support in the main android kotlin module
yet.

# Future developments / Not yet implemented in this PR

- Baseline management
- Updating previews (this can be done already, but with a crude
overwrite)
- More examples and more screen sizes and UI configurations
- Discover classes and methods automatically

---------

Co-authored-by: irodotos <irodotos@vaslabs.io>