Fix `transitiveCoursierProjects` cause OOM on large build (#4452) (#4454)
Resolve #4451 

It seems that `transitiveCoursierProjects` is populated with repetitive
transitive modules, that's why it only causes problem for large build.

Please open all PRs as drafts and ensure that your fork of Mill has 
`settings/actions` / `Allow all actions and reusable workflows` enabled
to run CI on
your own fork of the Mill repo. Only once CI passes mark the PR as
`Ready for review`
and CI will run on the main Mill repo before we merge it.

Co-authored-by: Mai Huy Hoàng <hoangmaihuy@users.noreply.github.com>