Initially copy and adapt some code to support `mill init` from an existing SBT project from its corresponding Gradle support
`SbtBuildGenMain.scala` and `BuildGenTests.scala` are not fully adapted yet, and the unadapted code is temporarily commented out.

An initial SBT plugin `ProjectTreePlugin` is added but not fully implemented yet. Only a task that writes some content to a file is added to test whether this approach works.

Exception in `./mill main.init.sbt.compile`:
```text
[1486] java.io.IOException: Scala signature package has wrong version
[1486]  expected: 5.0
[1486]  found: 5.2 in package.class
```

Corresponding code for Maven and Gradle is slightly improved BTW.