Introduce `mill.internal` package to hold Mill utility classes not intended for public use (#4507)
This PR pulls out a bunch of classes from `mill.api` and `mill.util`
that aren't meant for public use and moves them into `mill.internal`. We
also try to assign proper meanings to `mill.api` and `mill.util` that
were previously just grab-bags of random helpers

The meaning of the folders is now:

- `mill.api`: things that form the shared API between mill "core"
`define`/`eval`/`resolve`, and user-land Mill code
- `mill.internal`: things that are only used by Mill core code
- `mill.util`: things that are only used by user-land Mill code

This lets us clearly demarcate what is meant to be exposed to user code
and what isn't, which was very hard to see when we use `private`
modifiers on individual classes in `mill.api` and `mill.util`. Most of
the stuff in `mill.internal` is not related to any user-land APIs at all
- not even the `*lib` language toolchains - and truly is internal to
Mill's internal functionality.

`internal` is marked as a `MillPublishScalaModule` rather than
`MillStableScalaModule`: this lets us mark them as unstable and clearly
not meant to be relied upon by user code, while also leaving them
exposed so if anyone was relying on them before they can continue doing
so (for the moment) to aid in the migration.

We can continue to use `private` on individual classes or members where
appropriate for fine-grained visibility control; this just gives us a
big coarse-grained way of marking things as `internal`

The `main/` folder in Mill also probably should be re-organized along
these lines, with a split between `core` and non-`core` code. That can
come in a separate PR

This split can also serve the purpose of
https://github.com/com-lihaoyi/mill/pull/3260, where `mill.api` forms
the boundary between user-land code resolved by Coursier and Mill-Core
code bundled in the assembly. But that can also come later