Cleanup `mill.define.Module` and `mill.define.Ctx` (#4518)
* Bundle the necessary things we want to propagate together under
`Ctx.Nested`, rather than passing a loose bunch of implicits. This
reduces the amount of implicits we have flying around, and lets us
easily mark fields on `Ctx.Nested` as `private` where possible so we can
add or remove them

* Standardize naming convention of module members as `module*`:
`moduleDirectChildren`, `moduleCtx`, etc. Fixes
https://github.com/com-lihaoyi/mill/issues/4111 by renaming
`millSourcePath` to `moduleDir`

* Clean up `mill.define.Ctx`, removing a bunch of the special case
semantics around `RootModule`/`SubfolderModule`

* Add some basic unit tests in
`core/define/test/src/mill/define/ModuleTests.scala` to assert on the
module instantiation and context propagation, without the complexity of
downstream tests in `exec` or `resolve`

* Introduced `mill.define.internal` package, which is stuff in
`mill.define` that isn't really intended for direct use. Some of it is
`private` but needs to be in this module, some of it can't be `private`
but isn't really meant to be used directly.