Fix launch time regression in Scala 3 upgrade (#4526)
Fixes https://github.com/com-lihaoyi/mill/issues/4498

Turns out we were instantiating a new Classloader to hold the Scala 3
compiler worker every run, which is expensive, even though the worker is
the same every time. This PR moves the compile worker to a global `lazy
val` so it can be reused.

Also replaced `hexArray` with a implementation that avoids
`String.format` and `BigInteger`, because they were appearing in the
Jprofiler profiles, and replaces a `sha1Hash` call with `md5hex` since
the callsite doesn't require a cryptographic hash

Tested manually, makes warm launch times drop back to similar to 0.12.7

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>