Add option to use the embeddable compiler in `KotlinModule` (#4557)
Adds a new `kotlinCompilerEmbeddable` task, which switches the used
compiler toolchain to use the alternative "embeddable" jars.

This is in preparation for https://github.com/com-lihaoyi/mill/pull/4485
.

In summary, these changes are needed to avoid classpath conflicts
between the compiler and the compiler plugins (as it happens
[here](https://github.com/com-lihaoyi/mill/pull/4485) with the
dagger-compiler , which is used in Android DI setups as well (with the
addition of Hilt).

Pull request: https://github.com/com-lihaoyi/mill/pull/4557