Make `--no-server` mode also terminate on removal of `processId` file, use it in integration test cleanup  (#4587)
This attempts to fix some of the residual flakiness encountered in
https://github.com/com-lihaoyi/mill/pull/4570

Somehow the recursive process termination we pulled in from upstream
https://github.com/com-lihaoyi/os-lib/pull/359 only works some of the
time, and other times the processes get leaked. This PR extends the
`serverId`-file-deleted shutdown logic we already used for client-server
mode and enables it for `--no-server` mode as well. This lets us put an
additional guardrail in our `IntegrationTester#close` to delete any such
files, to try and force any Mill processes to terminate.

Added an additional integration test to exercise this behavior

Also fixed a bug in `ExampleTester` not honoring the `clientServerMode`
flag, and update `testkit.test` to assert on those behaviors