feat: RootModule and Cross can be applied to the same module (#4593)
Implements #3693 

These seem to be the minimal changes to allow RootModule and Cross to be
applied to the same module.
it was only necessary to make `Cross` a trait. I did try making
`RootModule` also a trait but it proved a little trickier due to it
usign the class constructor for `BaseModule` that it extends.

The main issue I ran into that caused me to have to remove the implicit
parameter on `Cross` was the Ctx implicit was not being filled in
properly and i was seeing `Modules and Tasks can only be defined within
a mill Module` from the dummy being used. There may have been a way
around this in the `CrossMacros`, but I couldn't see it, and this
solution appears to work, at least for the examples shown in the
integration tests added.