Provide default JVM version to be used by Mill server process, when `.mill-jvm-version` and system `java` are not provided (#4597)
This PR makes Mill's server use a fixed JVM version by default in the
case where no other Java version is available:

1. No `.mill-jvm-version` file present
2. No `java` command available environmentally
3. The Mill client has no `java.home` set (i.e. it is a graal native
image)

This makes running Mill on a clean environment without a JVM installed
work by default. Previously, it either needed you to define a
`.mill-jvm-version` or install a JVM, either of which is a manual step
we now can now avoid. Now it instead uses the version of Java configured
in Mill's own `.mill-jvm-version` file

We still allow any system `java` command to take precedence over the
default, since that's what most people would expect when using the JVM.

Added an integration test in CI `no-jvm-bootstrap` that runs Mill with
the native launcher with no `.mill-jvm-version` file and no system
`java` installed to verify that it falls back to the Mill default JVM
version. Needed to add some hacky test hooks to make sure we exercise
the code path in question