Android/publishing aar  (#4583)
## Motivation

As part of Mill’s Android integration, we want to create and publish
Android libraries so that multiple apps can be reused them. Initially,
we aim to publish AAR files locally, and then leverage them in Android
apps built with Mill.

## Provider in this PR

- Ability to build an AAR


![image](https://github.com/user-attachments/assets/64be0d2f-6758-4295-9b69-ce71de4e4ce4)

- Local publishing of AARs so projects can consume them via
~/.m2/repository


![image](https://github.com/user-attachments/assets/80704ca4-e5c6-45c2-9ed7-d7ab24ab5023)

- Use the local library to my Android app


![image](https://github.com/user-attachments/assets/416689d6-856d-4d28-bb4f-87f6c9a370c6)

## Refactoring

AndroidAppModule and AndroidLibModule had a lot of common functionality
so I created a separated AndroidModule that both app and lib module
extend. I did the same with AndroidKotlinModule

## Future Work

- Integration testing for both libraries and apps that uses the
libraries
- Publish to remote central repo