[WIP] Make `Evaluator` only available to `Task.Command(exclusive = true)` (#3717)
This is a breaking change that tightens things up a bit following the
introduction of `Task.Command(exclusive = true)`: now that we run
commands in parallel, running multiple evaluator commands in parallel is
dangerous, so they should be limited to `exclusive` commands

It's another step toward https://github.com/com-lihaoyi/mill/issues/502,
by declaring that `exclusive` commands are different from normal
commands, and thus need special treatment

Had to jump through some hoops to do this without breaking bincompat:

* We inject the Evaluator parameter during the resolution phase, before
evaluation starts,
* At that time, we do not yet know whether a command is exclusive or
not.
* The solution is to instead pass a `ProxyEvaluator` that exposes all
the same API, but looks up the current `Evaluator.currentEvaluator`
on-demand
* `Evaluator.currentEvaluator` that only exists within exclusive
commands, so we can fail with a nice error message if someone tries to
use the evaluator inside a non-exclusive command
* We cannot remove other usages of `Evaluator.currentEvaluator`, 
* e.g. in the `mill.main.Tasks.TokensReader` actually needs to resolve
the tasks early on and return a `Seq`, which we cannot easily stub out.
* But that's probably OK, because it doesn't expose the entire
`Evaluator`, and the tasks themselves are immutable and probably safe to
use concurrency