Integrate `os.Checker` to sandbox filesystem IO (#4635)
Fixes https://github.com/com-lihaoyi/mill/issues/3746

There are three main points at which user code will run

1. `RootModule` instantiation in `MillBuildBootstrap`
2. `resolve` phase, which instantiates module objects and tasks
* During `resolve`, we do not expect code to write to the filesystem at
all
3. `execute` phase, which executes task bodies in topological order. 
* Here we allow writes to each task's `.dest` folder and the `.dest`
folder of upstream worker tasks.
* Commands are run with sandboxing disabled, since those are meant for
performing non-hermetic side effects

This PR adds `os.checker.withValue` blocks around each of these points.
to check for invalid writes going through the `os.*` methods. For areas
that we need to disable these guardrails, we use `os.Checker.Nop`, which
is also exposed to users if they want to disable the guardrails for
whatever reason (e.g. we do this in the `example/thirdparty/acyclic`
suite).

The filesystem sandbox is not meant to be airtight, but rather are just
guardrails to nudge people in the right direction

For now this PR focuses on limiting writes, which are comparatively
simple. Limiting reads would take more thought and can come in a follow
up

Covered by a small integration test to exercise the three failure paths