Fix curl check in Windows wrapper script (#4648)
The curl check in the `mill.bat` wrapper script is actually broken:

```batch
if not exist "%MILL%" (
    ...
    where /Q curl
    if %ERRORLEVEL% EQU 0 (
        curl -f -L "!DOWNLOAD_URL!" -o "!DOWNLOAD_FILE!"
    )
```

The `%ERRORLEVEL%` is inside of nested `if`, so it will be the value of
`ERRORLEVEL` when `if not exist "%MILL%"` is executed, not the result of
`where /Q curl`. It should be replaced with `!ERRORLEVEL!` to get the
correct result.

Some minor indentation fixes are also included.

Pull request: https://github.com/com-lihaoyi/mill/pull/4648