Remove PrintLogger (#4661)
Part of https://github.com/com-lihaoyi/mill/issues/3603

We've been using the new multi-line `PromptLogger` for a while since
0.12.0 was released back in Aug 2024, and it's worked great. So we can
remove the old implementation for 0.13.0 due in mid 2025, and deprecate
and hide the `--remove-prompt` flag that was introduced to flip back to
the old implementation.

It turns out that we were still using the old `PrintLogger` in a bunch
of places - unit tests, testrunner, etc. - and I had to replace those
use sites either with `PromptLogger` or with just `println`

Also made some adjustments to `enableTicker` handling to keep it
internal within `Logger` so it is more consistently applied, and fixed
the logging of skipped tasks in `--keep-going`